+++ objc2-foundation/src/array.rs
@@ -14,7 +14,7 @@ use super::{
     NSRange,
 };
 
-object! {
+__inner_object! {
     /// TODO
     ///
     /// You can have a `Id<NSArray<T, Owned>, Owned>`, which allows mutable access
@@ -43,7 +43,7 @@ unsafe impl<T: Sync + Send> Send for NSArray<T, Shared> {}
 unsafe impl<T: Sync> Sync for NSArray<T, Owned> {}
 unsafe impl<T: Send> Send for NSArray<T, Owned> {}
 
-object! {
+__inner_object! {
     // TODO: Ensure that this deref to NSArray is safe!
     // This "inherits" NSArray, and has the same `Send`/`Sync` impls as that.
     #[derive(Debug, PartialEq, Eq, Hash)]